+++ CHANGELOG.md
@@ -10,9 +10,9 @@ CHANGELOG
 
 * 0.3.0 (2013-xx-xx)
 
- * BC: Require hostname and do Origin HTTP header check against it by default, helping prevent CSRF attacks
- * Added HTTP Router component to allowing for a single Ratchet server to handle multiple apps
- * BC: Decoupled HTTP from WebSocket component
+ * Added Symfony/2.2 based HTTP Router component to allowing for a single Ratchet server to handle multiple apps -> Ratchet\Http\Router
+ * BC: Decoupled HTTP from WebSocket component -> Ratchet\Http\HttpServer
+ * Updated dependency to React/0.3
 
 * 0.2.5 (2013-04-01)